# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BioCypher Components Registry API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/registrations": {
      "post": {
        "tags": [
          "registrations"
        ],
        "summary": "Submit an adapter registration",
        "description": "Store one adapter repository as a tracked registration request. The repository may be a local path or remote URL and should contain a root-level croissant.jsonld file. Processing happens later through the process endpoint or a registry refresh.",
        "operationId": "create_registration_api_v1_registrations_post",
        "parameters": [
          {
            "name": "bcr_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bcr Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "registrations"
        ],
        "summary": "List registrations",
        "description": "Return active registration summary rows. This list is intentionally metadata-light; use the registration detail endpoint for processing details and metadata payloads.",
        "operationId": "list_registrations_api_v1_registrations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/croissant-file-present-check": {
      "get": {
        "tags": [
          "registrations"
        ],
        "summary": "Check registration Croissant file presence",
        "description": "Return whether a remote repository exposes root-level croissant.jsonld.",
        "operationId": "check_registration_croissant_file_presence_api_v1_registrations_croissant_file_present_check_get",
        "parameters": [
          {
            "name": "repository_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Repository Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationCroissantFilePresentCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registration_id}/events": {
      "get": {
        "tags": [
          "registrations"
        ],
        "summary": "List registration events",
        "description": "Return the event history recorded while submitting, processing, refreshing, or revalidating one registration.",
        "operationId": "list_registration_events_api_v1_registrations__registration_id__events_get",
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationEventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registration_id}": {
      "get": {
        "tags": [
          "registrations"
        ],
        "summary": "Get registration detail",
        "description": "Return one operator-facing registration detail record. This endpoint may include discovered metadata, metadata path, validation errors, and canonical uniqueness details.",
        "operationId": "get_registration_api_v1_registrations__registration_id__get",
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registration_id}/process": {
      "post": {
        "tags": [
          "registrations"
        ],
        "summary": "Process one registration",
        "description": "Discover the submitted repository metadata, validate the adapter and embedded datasets, update registration status, and create or reuse a canonical registry entry when valid.",
        "operationId": "process_registration_api_v1_registrations__registration_id__process_post",
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationProcessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registration_id}/revalidate": {
      "post": {
        "tags": [
          "registrations"
        ],
        "summary": "Revalidate one registration",
        "description": "Reprocess one registration whose current status is INVALID or whose latest event is FETCH_FAILED.",
        "operationId": "revalidate_registration_route_api_v1_registrations__registration_id__revalidate_post",
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationRevalidateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegistrationCreateRequest": {
        "properties": {
          "adapter_name": {
            "type": "string",
            "minLength": 1,
            "title": "Adapter Name",
            "description": "Human-readable adapter name supplied by the maintainer."
          },
          "repository_location": {
            "type": "string",
            "minLength": 1,
            "title": "Repository Location",
            "description": "Local repository path or supported remote repository URL containing a root-level croissant.jsonld file."
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email",
            "description": "Optional maintainer contact email for follow-up."
          },
          "license_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Value",
            "description": "Optional submitted adapter license text."
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi",
            "description": "Optional submitted DOI text."
          },
          "cff_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cff Url",
            "description": "Optional submitted Citation File Format URL."
          }
        },
        "type": "object",
        "required": [
          "adapter_name",
          "repository_location"
        ],
        "title": "RegistrationCreateRequest",
        "description": "Request body for creating an adapter registration.",
        "example": {
          "adapter_name": "CollecTRI Adapter",
          "cff_url": "https://gitlab.example.org/biocypher/collectri/-/raw/main/CITATION.cff",
          "contact_email": "maintainer@example.org",
          "doi": "10.5281/zenodo.1234567",
          "license_value": "MIT",
          "repository_location": "https://gitlab.example.org/biocypher/collectri"
        }
      },
      "RegistrationCreateResponse": {
        "properties": {
          "registration_id": {
            "type": "string",
            "title": "Registration Id"
          },
          "adapter_name": {
            "type": "string",
            "title": "Adapter Name"
          },
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id"
          },
          "repository_location": {
            "type": "string",
            "title": "Repository Location"
          },
          "repository_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "local",
                  "remote"
                ]
              },
              {
                "type": "string"
              }
            ],
            "title": "Repository Kind"
          },
          "status": {
            "$ref": "#/components/schemas/RegistrationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "license_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Value"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "cff_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cff Url"
          },
          "submitted_by_github_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By Github Login"
          }
        },
        "type": "object",
        "required": [
          "registration_id",
          "adapter_name",
          "adapter_id",
          "repository_location",
          "repository_kind",
          "status",
          "created_at"
        ],
        "title": "RegistrationCreateResponse",
        "description": "Response returned after creating an adapter registration."
      },
      "RegistrationCroissantFilePresentCheckResponse": {
        "properties": {
          "has_croissant_file": {
            "type": "boolean",
            "title": "Has Croissant File"
          }
        },
        "type": "object",
        "required": [
          "has_croissant_file"
        ],
        "title": "RegistrationCroissantFilePresentCheckResponse",
        "description": "Response model for a remote croissant.jsonld presence check."
      },
      "RegistrationDetailResponse": {
        "properties": {
          "registration_id": {
            "type": "string",
            "title": "Registration Id"
          },
          "adapter_name": {
            "type": "string",
            "title": "Adapter Name"
          },
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id"
          },
          "repository_location": {
            "type": "string",
            "title": "Repository Location"
          },
          "repository_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "local",
                  "remote"
                ]
              },
              {
                "type": "string"
              }
            ],
            "title": "Repository Kind"
          },
          "status": {
            "$ref": "#/components/schemas/RegistrationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "license_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Value"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "cff_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cff Url"
          },
          "submitted_by_github_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By Github Login"
          },
          "metadata_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Path"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "profile_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "uniqueness_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uniqueness Key"
          },
          "validation_errors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Errors"
          }
        },
        "type": "object",
        "required": [
          "registration_id",
          "adapter_name",
          "adapter_id",
          "repository_location",
          "repository_kind",
          "status",
          "created_at"
        ],
        "title": "RegistrationDetailResponse",
        "description": "Detailed response for one stored registration."
      },
      "RegistrationEventListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegistrationEventResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "RegistrationEventListResponse",
        "description": "Response model for a registration event history list."
      },
...
      "RegistrationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegistrationListItemResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "RegistrationListResponse",
        "description": "Response model for a registration list."
      },
      "RegistrationProcessResponse": {
        "properties": {
          "registration_id": {
            "type": "string",
            "title": "Registration Id"
          },
          "adapter_name": {
            "type": "string",
            "title": "Adapter Name"
          },
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id"
          },
          "repository_location": {
            "type": "string",
            "title": "Repository Location"
          },
          "repository_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "local",
                  "remote"
                ]
              },
              {
                "type": "string"
              }
            ],
            "title": "Repository Kind"
          },
          "status": {
            "$ref": "#/components/schemas/RegistrationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "license_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Value"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "cff_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cff Url"
          },
          "submitted_by_github_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By Github Login"
          },
          "metadata_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Path"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "profile_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "uniqueness_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uniqueness Key"
          },
          "validation_errors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Errors"
          }
        },
        "type": "object",
        "required": [
          "registration_id",
          "adapter_name",
          "adapter_id",
          "repository_location",
          "repository_kind",
          "status",
          "created_at"
        ],
        "title": "RegistrationProcessResponse",
        "description": "Response returned after processing one registration."
      },
...
      "RegistrationStatus": {
        "type": "string",
        "enum": [
          "SUBMITTED",
          "VALID",
          "INVALID"
        ],
        "title": "RegistrationStatus",
        "description": "Tracked states for a stored registration request."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}