# src/api/routers/registrations.py
"""Registration workflow routes."""

from __future__ import annotations

from typing import Annotated

from fastapi import APIRouter, Depends, HTTPException, Query, status

from src.api.dependencies import get_optional_auth_session, get_registration_store
from src.api.errors import (
    registration_not_found_http_error,
    registration_processing_http_error,
    registration_submission_http_error,
)
from src.api.schemas.registrations import (
    RegistrationCreateRequest,
    RegistrationCreateResponse,
    RegistrationCroissantFilePresentCheckResponse,
    RegistrationDetailResponse,
    RegistrationEventListResponse,
    RegistrationEventResponse,
    RegistrationListItemResponse,
    RegistrationListResponse,
    RegistrationProcessResponse,
    RegistrationRevalidateResponse,
)
from src.core.auth.models import AuthSession
from src.core.registration.service import (
    finish_registration,
    revalidate_registration,
    submit_registration,
)
from src.core.registration.store import RegistrationStore
from src.core.shared.files import remote_metadata_exists


router = APIRouter()

RegistrationStoreDep = Annotated[RegistrationStore, Depends(get_registration_store)]
OptionalAuthSessionDep = Annotated[AuthSession | None, Depends(get_optional_auth_session)]


# ===========================================================
# Registration Routes
# ===========================================================


@router.post(
    "/registrations",
    status_code=status.HTTP_201_CREATED,
    summary="Submit an adapter registration",
    description=(
        "Store one adapter repository as a tracked registration request. "
        "The repository may be a local path or remote URL and should "
        "contain a root-level croissant.jsonld file. Processing happens later "
        "through the process endpoint or a registry refresh."
    ),
)
def create_registration(
    payload: RegistrationCreateRequest,
    store: RegistrationStoreDep,
    auth_session: OptionalAuthSessionDep,
) -> RegistrationCreateResponse:
    """Create and persist a submitted adapter registration."""
    try:
        registration = submit_registration(
            adapter_name=payload.adapter_name,
            repository_location=payload.repository_location,
            store=store,
            contact_email=payload.contact_email,
            license_value=payload.license_value,
            doi=payload.doi,
            cff_url=payload.cff_url,
            submitted_by_github_login=(
                auth_session.github_login if auth_session is not None else None
            ),
        )
    except Exception as exc:  # noqa: BLE001
        raise registration_submission_http_error(exc) from exc

    return RegistrationCreateResponse.from_stored(registration)


@router.get(
    "/registrations",
    summary="List registrations",
    description=(
        "Return active registration summary rows. This list is intentionally "
        "metadata-light; use the registration detail endpoint for processing "
        "details and metadata payloads."
    ),
)
def list_registrations(
    store: RegistrationStoreDep,
) -> RegistrationListResponse:
    """Return active stored adapter registrations."""
    registrations = store.list_active_registrations()
    items = [
        RegistrationListItemResponse.from_stored(registration)
        for registration in registrations
    ]

    return RegistrationListResponse(items=items)


@router.get(
    "/registrations/croissant-file-present-check",
    response_model=RegistrationCroissantFilePresentCheckResponse,
    summary="Check registration Croissant file presence",
    description="Return whether a remote repository exposes root-level croissant.jsonld.",
)
def check_registration_croissant_file_presence(
    repository_url: str = Query(..., min_length=1),
) -> RegistrationCroissantFilePresentCheckResponse:
    """Return whether the submitted repository exposes Croissant metadata."""
    try:
        has_croissant_file = remote_metadata_exists(repository_url)
    except Exception as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="We could not check the repository url you provided (we tried to check if a raw croissant.jsonld file could be found by our system) Is your file hosted on GitHub or a Gitlab instance?",
        ) from exc
    return RegistrationCroissantFilePresentCheckResponse(
        has_croissant_file=has_croissant_file
    )
# previously this happened on the frontend but with Gitlab + Github support it became quite messy.
# Adding Gitlab support has added about 30 lines to the project, but we keep it simpler conceptually by only hcecking on the backend.


@router.get(
    "/registrations/{registration_id}/events",
    summary="List registration events",
    description=(
        "Return the event history recorded while submitting, processing, "
        "refreshing, or revalidating one registration."
    ),
)
def list_registration_events(
    registration_id: str,
    store: RegistrationStoreDep,
) -> RegistrationEventListResponse:
    """Return the event history for one stored adapter registration."""
    registration = store.get_registration(registration_id)
    if registration is None:
        raise registration_not_found_http_error(registration_id)

    events = store.list_registration_events(registration_id)
    return RegistrationEventListResponse(
        items=[RegistrationEventResponse.from_event(event) for event in events]
    )


@router.get(
    "/registrations/{registration_id}",
    summary="Get registration detail",
    description=(
        "Return one operator-facing registration detail record. This endpoint "
        "may include discovered metadata, metadata path, validation errors, and "
        "canonical uniqueness details."
    ),
)
def get_registration(
    registration_id: str,
    store: RegistrationStoreDep,
) -> RegistrationDetailResponse:
    """Return one stored adapter registration by identifier."""
    registration = store.get_registration(registration_id)
    if registration is None:
        raise registration_not_found_http_error(registration_id)

    return RegistrationDetailResponse.from_stored(registration)


@router.post(
    "/registrations/{registration_id}/process",
    summary="Process one registration",
    description=(
        "Discover the submitted repository metadata, validate the adapter and "
        "embedded datasets, update registration status, and create or reuse a "
        "canonical registry entry when valid."
    ),
)
def process_registration(
    registration_id: str,
    store: RegistrationStoreDep,
) -> RegistrationProcessResponse:
    """Discover, validate, and persist one stored registration result."""
    try:
        registration = finish_registration(
            registration_id=registration_id,
            store=store,
        )
    except Exception as exc:  # noqa: BLE001
        raise registration_processing_http_error(registration_id, exc) from exc

    return RegistrationProcessResponse.from_stored(registration)


@router.post(
    "/registrations/{registration_id}/revalidate",
    summary="Revalidate one registration",
    description=(
        "Reprocess one registration whose current status is INVALID or whose "
        "latest event is FETCH_FAILED."
    ),
)
def revalidate_registration_route(
    registration_id: str,
    store: RegistrationStoreDep,
) -> RegistrationRevalidateResponse:
    """Reprocess one previously invalid or fetch-failed registration."""
    try:
        registration = revalidate_registration(
            registration_id=registration_id,
            store=store,
        )
    except Exception as exc:  # noqa: BLE001
        raise registration_processing_http_error(registration_id, exc) from exc

    return RegistrationRevalidateResponse.from_stored(registration)